requirements.txt
torch==1.7.0
PyYAML==5.3.1
numpy==1.19.2
h5py==2.10.0
opencv-python==4.4.0.44
pre-commit
matplotlib==3.3.2
progress
mlflow